+++ src/main/kotlin/systems/fehn/intellijdirenv/listeners/MyProjectManagerListener.kt
@@ -8,35 +8,32 @@ import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.ProjectManagerListener
 import systems.fehn.intellijdirenv.MyBundle
+import systems.fehn.intellijdirenv.notificationGroup
 import systems.fehn.intellijdirenv.services.DirenvProjectService
-import systems.fehn.intellijdirenv.services.DirenvService
 
 internal class MyProjectManagerListener : ProjectManagerListener {
     private val logger by lazy { logger<MyProjectManagerListener>() }
-    private val direnvService by lazy { service<DirenvService>() }
 
     override fun projectOpened(project: Project) {
         logger.trace("Opened project ${project.name}")
 
-        direnvService.direnvExecutable?.let { executable ->
-            val projectService = project.service<DirenvProjectService>()
+        val projectService = project.service<DirenvProjectService>()
 
-            if (projectService.hasEnvrcFile()) {
-                val notification = direnvService.notificationGroup
-                    .createNotification(
-                        MyBundle.message("envrcFileFound"),
-                        NotificationType.INFORMATION,
-                    )
-                    .addAction(
-                        NotificationAction.create(MyBundle.message("importDirenvAction")) { _, notification ->
-                            notification.hideBalloon()
+        if (projectService.hasEnvrcFile()) {
+            val notification = notificationGroup
+                .createNotification(
+                    MyBundle.message("envrcFileFound"),
+                    NotificationType.INFORMATION,
+                )
+                .addAction(
+                    NotificationAction.create(MyBundle.message("importDirenvAction")) { _, notification ->
+                        notification.hideBalloon()
 
-                            projectService.importDirenv(executable)
-                        },
-                    )
+                        projectService.importDirenv()
+                    },
+                )
 
-                Notifications.Bus.notify(notification, project)
-            }
+            Notifications.Bus.notify(notification, project)
         }
     }
 }