+++ src/main/kotlin/systems/fehn/intellijdirenv/utils.kt
@@ -1,5 +1,7 @@
 package systems.fehn.intellijdirenv
 
+import com.intellij.notification.NotificationGroup
+
 fun <T : Any> T?.switchNull(
     onNull: (() -> Unit)? = null,
     onNonNull: ((T) -> Unit)? = null,
@@ -10,3 +12,5 @@ fun <T : Any> T?.switchNull(
         onNonNull?.invoke(thing)
     }
 }
+
+val notificationGroup = NotificationGroup.balloonGroup("Direnv")